Moved modules and service to plugins directory
