+++ IOMC/RandomEngine/plugins/BuildFile.xml
@@ -1,6 +1,6 @@
 <use name="FWCore/ServiceRegistry"/>
 <use name="FWCore/Framework"/>
 <use name="IOMC/RandomEngine"/>
-<library file="Module.cc" name="IOMCRandomEnginePlugins">
+<library file="*.cc" name="IOMCRandomEnginePlugins">
   <flags EDM_PLUGIN="1"/>
 </library>
